Compartmentalize GPU and graphic apps into VMs
This commit introduces several new entities such as GPU VM
and Application VMs. To use those, you need to properly set
up devices passthroug first. Refer to the project documentation
on how to achieve that.

Lenovo X1 Carbon Gen11 is introduced as a new target and the
reference device for implementing compartmentalization.

It is still possible to build standard "mainline" ghaf with
weston and demo apps working on host.

* User apps are now added into the scope of nixpkgs;
* minimal.nix module is disabled since it introduced some
weird metadata changes.

Co-authored-by: Berk Arslan <berk.arslan@unikie.com>
Co-authored-by: Mika Tammi <mika.tammi@unikie.com>
Co-authored-by: Nikita Bazulin <nikita.bazulin@unikie.com>
Co-authored-by: Vadim Likholetov <vadim.likholetov@unikie.com>
Co-authored-by: Valentin Kharin <valentin.kharin@unikie.com>
Co-authored-by: Yuriy Nesterov <yuriy.nesterov@unikie.com>

Signed-off-by: Ivan Nikolaenko <ivan.nikolaenko@unikie.com>