Initial move to flake-parts
This is the start of refactoring Ghaf to be more modular.

This patch introduces the use f flake-parts to aid in this
regard, to help passing state and configurations between
modules.

Also introduced are some linting checks
1. reuse - licenses
2. deadnix - unused code
3. statix - avoid anti-patterns
4. alejandra - code formatting

```nix flake check``` is adjusted so that it works again.

the default ´devshell` is now configured providing access to all
needed tools and also introduces 2 new tools for aid in the CI process.

1. nix-build-all - build everything (like check flake)
2. nix-fast-build - do more in parallel

To use the ``devshell`` you can use either.

``direnv allow`` - to always open a devshell for the directory

or

``nix develop`` - to perform on demand devshell invocation.

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>