Add Waybar configuration
Waybar configuration modified based on the default bar configuration and launchers added.
Pulseaudio is added to guivm.
Nmcli is used to get the network connection data in wifi-signal-strength script and nm-launcher connection structure is copied.
Weston and waybar launchers are unified and name section added.

Nerdfonts FiraCode is added.
All the style and configuration files are created in weston.config.nix.

Compositor and waypipe display name must be overriden as in comments in guivm.nix to use waybar properly.

Signed-off-by: Berk Arslan <berk.arslan@unikie.com>