Add an XDG handler for PDF files to the chromium-vm
The XDG handler script is executed by Chromium when the user attempts
to open a PDF file. The script sends the PDF file path to the GUIVM
over a TCP connection. The GUIVM copies the file from chromium-vm to a
temporary location in the zathura-vm and opens it there. Once the viewer
is closed, the file is deleted from the zathura-vm.

Signed-off-by: Yuri Nesterov <yuriy.nesterov@unikie.com>