Ignore patch file number in shouldOverride
On checking if the systemd timesyncd patch is applied nixpkgs
upstream, do not require a match on the sequential patch file number.
Patches can be applied on different order, or the total number of
patches might change over time, so it's incorrect to assume the
timesyncd patch would always have the sequential order of '0020'.

Signed-off-by: Henri Rosten <henri.rosten@unikie.com>