labwc: add launcher and other improvements
This patch adds nwg-drawer with the launchers moved in the drawer,
instead of appending them on the top bar.

This allows us to define more application launchers in the future, and
even categorise them per virtual machine.

It also makes the status bars (top left) more readable and
user-friendly, showing the SSID by default instead of IP, and showing
the full date and time.

It also moves the definition for the keyboard layout into the systemd
environment variables, and sets a sane default 'us,fi'.

This doesn't affect Weston launchers, launchers on Weston remains and
functions the same way without the drawer.

Signed-off-by: Humaid Alqasimi <humaid.alqassimi@tii.ae>