Orin: UARTI passthrough to Net-VM
Having a USB debug port connected to the Net-VM is
useful when the SSH connection is not working or correctly
configured. This implementation enables access from the
Nvidia Orin (AGX/NX) debug micro USB port (ttyACM1) to
the Net-VM console.

- Use the hardware.deviceTree nix module for the host device
tree overlay in the UARTI passthrough to net-vm.
- Added documentation of how to use hardware.deviceTree
overlays in ghaf for passthrough devices.
- Fixed documentation according to received feedback.
- Fixed problems with device tree differences on jetpack 35.4.1

Signed-off-by: Juan Pablo Ruiz <juanpablo.ruiz@unikie.com>