gpio-vm starts up with an unique Device Tree
gpio-vm's device tree tegra234-p3701-0000-gpio-passthrough.dtb declares
the vda device needed for passthrough.

gpio-vm goes into emergency mode because it cannot open AMA0 as a
console. Uarta passthrough is disabled because of conflicts with GPIO
passthrough. We need to disable the QEMU console. Somewhere in the code
console=ttyAMA0 is set for QEMU. It is not obvious where.