log-vm: starting systemd-journal-remote service in log-vm
Without this patch if ghaf-host tries to do `systemd-journal-upload`
there will be following error "Failed to connect to 192.168.101.66
port 19532 after 0 ms: Couldn't connect to server".
As there will be no listener running in log-vm to accept logs.

With this patch we add listener in log-vm on 19532 port to accept logs.
Logs will be stored at "/var/log/journal/remote" in log-vm.
`ghaf-host` logs are getting stored in log-vm successfully.

Signed-off-by: Vunny Sodhi <vunny.sodhi@unikie.com>