x1-installer: add checks in installer script, other improvements
The image path is now set as an environment variable, so the script
derivation could be reused. Also the installer derivation is created
using `writeShellApplication` which will test against shellcheck.

Also added message on running the installer. The installer now only
shows devices and the vendor/name, and doesn't include the partitions.

The installer also has proper confirmation prompts, and also warn the
user if they are trying to install Ghaf on a removable device.

Signed-off-by: Humaid Alqasimi <humaid.alqassimi@tii.ae>