Add make-checks to evaluate a derivation
This adds a new package to test the evaluation of all the outputs that
are produced from the flake.

This is a basic sanity check, it does not build all the targets and
packages as that is covered by `nix-fast-build` and requires even more
substantial resources and time to complete.

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>