store the screenshots from grim across reboots
In addition this patch adds the ability to select the capture area that
is used for the screenshots. When the screenshot is selected there is
Satty functionality that allows for some basic post capture editing in
the form of an overlay to add annotations such as, arrows, boxes, and
text.

The user can the save the file, or copy it to the clipboard. Once in the
clipboard it can be pasted into any of the other vms where there is an
app that will accept those images. This is the case for e.g. slack and
outlook.

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>