Add a calculator to the ui vm
The calculator is a basic app with no network requirements so adding it
to the ui vm will just make it commonly available.

The venerable gnome-calculator is the best simple calculator from the
range of calculators that have been tested. It allows for both basic
calculator view with a history of the calculations and it also offers
modes for advanced and programming modes.

As such it should cover all cases although it is still quite a plain
looking app it is powerful and useful.

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>