Open trusted browser link in normal browser
Add chromium extension to trusted browser that allows opening links and
pages in normal browser.

Extension adds "Open link in normal browser" -option to the right-click
pop-up menu for links in trusted browser and "Open page in trusted
browser" option for pop-up menu for rest of the page. Clicking the
extension icon will also open the current tab in normal browser.

The extension itself is a bash script that uses givc-cli to launch
the normal browser for the according URL.

Co-authored-by: Alexander Nikolaev <alexander.nikolaev@unikie.com>
Signed-off-by: Tero Tervala <tero.tervala@unikie.com>