Add dbus proxy via givc
- add dbus proxy with initial config for guivm to netvm/audiovm
- add blueman-mechanism service

General design:

GIVC provides a xdg-dbus-proxy module, running in the VM providing the service.
The module provides a policy filter, which is configured to allow communication with a client.
The xdg-dbus-proxy creates a local socket, which is proxied via givc/tcp. Policies can
be created by adding `debug = true;` to the dbus proxy module and/or the agents. The latter
will log all communications between the endpoints.

Client applications can connect to the local socket, provided by the givc
agent. Upon connection, a new grpc stream is initiated to the target vm,
which in turn connects to the local xdg-dbus-proxy.

Signed-off-by: Manuel Bluhm <manuel@ssrc.tii.ae>