zfs: Enable LUKS encryption on data and swap partitions
This patch create separate zfs data pool and swap
partitions with LUKS encryption. The password and
TPM PIN for LUKS encrypted partitions are provided by user
which will be later stored on TPM security chip and will
be decrypted automatically on the subsequent system boots
with need of user to enter TPM2 token PIN on every boot.

Signed-off-by: Vunny Sodhi <vunny.sodhi@unikie.com>