Fix Element location sharing
gpsd was not able to connect to USB gps device because debug serial
console was enabled for comms-vm. As gps device shows up as serial port,
the device was assigned to 'tty' group and gpsd could not access it, as
gpsd runs with effective gid of 'dialout' group.

Run gpsd with a control socket and add udev rules to allow hotplugging
of gps devices without additional manual restarting of services.

Also made sure that element-gps service is started after gpsd service.

Signed-off-by: Tero Tervala <tero.tervala@unikie.com>