audio: fix the audio jack
The diff below fixes issue (for now) basically it makes wireplumber
auto-select the output audio port. It also disables my script that
looped through the audio profiles at boot. It seems it is not needed
anymore. I think the reason for the constant changes with pipewire /
wireplumber is that they are in constant development and there is big
discussions going on should the audio system change the audio output
port when new device is connected or not.

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>
Co-authored-by: Jon Sahlberg <jon.sahlberg@unikie.com>