dell-latitude-7330-72: Add new build targets
This is needed as net-vm will fail to start as we cannot
have two same network devices which may enumerate on
different PCI bus on dell latitude 7330.
This is can be workaround for time
being until we find proper way to identify on which
PCI bus network controller is detected.
So adding debug and release build variants for
"dell-latitude-7330-72"

Signed-off-by: Vunny Sodhi <vunny.sodhi@unikie.com>