vhotplug: Enable type-c display for x86_64 variants
We need to pass usb class to vhotplug and opregion
to gui-vm so that it can detect type-c display.

As we are doing GPU pasthrough to `gui-vm` qemu
we need pass video BIOS table (VBT) info from
UEFI to `gui-vm` qemu for which qemu need to
access `x-igd-opregion` memory region,
hence enabling the same.

Signed-off-by: Vunny Sodhi <vunny.sodhi@unikie.com>