make-iso9660-image: copy intermediate steps to the derivation output
make-iso9660-image.sh is a nixpkgs script run when building a
system.build.isoImage derivation. It uses a combination of mksquashfs and
xorriso to create the final image.

Since signing the bootloader for secure boot involves unpacking the iso,
we output the intermediate commands used in make-iso9660-image so that
we can recreate an identical image after signing.