Alienware: Control display brightness
For platforms with Nvidia GPUs like Alienware,
controlling display brightness using the
intel_backlight interface is not possible.
To address this, we create a socket in the
GUI-VM and use the ACPI interface to write
to the socket. This socket is then utilized
by the host to update the display brightness
value.

Signed-off-by: Vunny Sodhi <vunny.sodhi@unikie.com>