microvm: storageVM encryption support for all VMs
This commit adds two methods to provide a TPM to a microvm:
1. passthrough of the host TPM or,
2. emulated TPM hosted in admin VM
The default configuration enables the passthrough for the system VMs and
the emulation for app VMs.

The storage VM has a new option: storagevm.encryption.enable. When this
is set, the persistent storage is backed by a disk image instead
of a virtiofs mapped directory. The disk image is encrypted inside the
VM using the TPM device available to the VM (through one of the 2
methods) and decrypted when the VM boots.

Signed-off-by: Hugo Ros <Hugo.Ros@tii.ae>