lenovo-x1/gen12: drop Intel MEI communication controller
Remove the Intel Management Engine Interface (MEI)
communication controller from the GPU PCI devices list.
The MEI controller may have different devices within
the same IOMMU group, which also need to be added
and can vary across different SKUs.Therefore, it is
best to drop it for now and add it back if required
in the future.

Signed-off-by: Vunny Sodhi <vunny.sodhi@unikie.com>