The lsp made me do it
nixd, through nixt and nixf highlights anti-patterns that should be
addressed. Mainly with shadowing functions between the different
namespaces, e.g. builtins and lib.

The general rule is that you should be explicit, if the function exists
in both scopes and not rely on inherit, to set the correct set.

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>