actions: accept the flake config for cache
If there are large changes in the package set, that is not included in
the default nixos cache. The eval and check jobs will fail continuously
in github action runners, due to running out of space. So utilize the
cache.

This of course requires that the build will run once to populate the
cache, then on subsequent runs actions will use these results and pass.

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>