iso: do not copy the system closure only the disk
The iso was containing the disk image that is to be flashed to the
target device along with the entire system closure. This is why the disk
image of 6.4 GB (system76 darter pro) was causing an iso of 13GB to be
created.

Now we ensure not to copy the closure just the disk image. There is
still room for improvement, but that will come in a follow up PR.

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>