ghaf-vms: similar to microvm cmdline tool
Adds functionality to host to query the status of the vms that are on
the system.

When using the storeOnDisk option, there is no link created to the
current system so the `microvm -l` will not work and just stops. Ensure
that it is possible to see at a glance which vm is actually running/stopped.

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>