(feat): Enable full disk encryption
This change modifies the disk partitioning scheme
to use `LVM on LUKS`, compared to the previous scheme
which used `LUKS on LVM`. The advantages of the new
scheme include reduced layout complexity and the
ability to enable full disk encryption. With this
approach, there is no need to encrypt different
partitions separately.

Signed-off-by: Vunny Sodhi <vunny.sodhi@unikie.com>